Spell Checker
Using the words from the csv files, this program places the possible letters into Tries and each node holds a word and letter representing the position. If given a word, the program will search through the tries to find it. If not found, it will give all words that are at most one letter off from the csv dictionary given.